Add logic for syncing restore with mongos
