+++ e2e-tests/functions
@@ -27,7 +27,6 @@ if oc get projects ; then
     esac
 fi
 
-
 create_namespace() {
     local namespace="$1"
     local skip_clean_namespace="$2"
@@ -191,6 +190,30 @@ wait_restore() {
     simple_data_check "${cluster}" "${2}" "$isSharded" "$cluster_pfx"
 }
 
+wait_deployment() {
+    local name=$1
+
+    sleep 10
+    set +o xtrace
+    retry=0
+    echo -n $name
+    until [ "$(kubectl_bin get deployment $name -o jsonpath='{.status.replicas}')" == "$(kubectl_bin get deployment $name -o jsonpath='{.status.readyReplicas}')"  ]; do
+        sleep 1
+        echo -n .
+        let retry+=1
+        if [ $retry -ge 360 ]; then
+            kubectl_bin logs $(get_operator_pod) \
+                | grep -v 'level=info' \
+                | grep -v 'level=debug' \
+                | grep -v 'Getting tasks for pod' \
+                | grep -v 'Getting pods from source' \
+                | tail -100
+            echo max retry count $retry reached. something went wrong with operator or kubernetes cluster
+            exit 1
+        fi
+    done
+}
+
 simple_data_check() {
     cluster_name=${1}
     let last_pod="$2-1" || :
@@ -198,6 +221,8 @@ simple_data_check() {
     local cluster_pfx=$4
 
     if [ $isSharded -eq 1 ]; then
+        wait_deployment "$cluster_name-mongos"
+        sleep 10
         compare_mongos_cmd "find" "myApp:myPass@$cluster$cluster_pfx.$namespace"
     else
         for i in $(seq 0 $last_pod); do
@@ -790,3 +815,4 @@ run_backup() {
 
     kubectl_bin apply -f $test_dir/conf/backup-$storage.yml
 }
+
