+++ pkg/controller/perconaservermongodb/backup.go
@@ -6,12 +6,13 @@ import (
 	"fmt"
 
 	batchv1b "k8s.io/api/batch/v1beta1"
-	"k8s.io/apimachinery/pkg/api/errors"
+	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/labels"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	api "github.com/percona/percona-server-mongodb-operator/pkg/apis/psmdb/v1"
 	"github.com/percona/percona-server-mongodb-operator/pkg/psmdb/backup"
+	"github.com/pkg/errors"
 )
 
 func (r *ReconcilePerconaServerMongoDB) reconcileBackupTasks(cr *api.PerconaServerMongoDB) error {
@@ -30,7 +31,7 @@ func (r *ReconcilePerconaServerMongoDB) reconcileBackupTasks(cr *api.PerconaServ
 			}
 
 			err = r.client.Create(context.TODO(), cjob)
-			if err != nil && errors.IsAlreadyExists(err) {
+			if err != nil && k8sErrors.IsAlreadyExists(err) {
 				err := r.client.Update(context.TODO(), cjob)
 				if err != nil {
 					return fmt.Errorf("update task %s: %v", task.Name, err)
@@ -71,7 +72,7 @@ func (r *ReconcilePerconaServerMongoDB) reconcileBackupTasks(cr *api.PerconaServ
 			}
 		} else {
 			err := r.client.Delete(context.TODO(), &t)
-			if err != nil && !errors.IsNotFound(err) {
+			if err != nil && !k8sErrors.IsNotFound(err) {
 				return fmt.Errorf("delete backup task %s: %v", t.Name, err)
 			}
 		}
@@ -143,4 +144,43 @@ func (h *minHeap) Pop() interface{} {
 	x := old[n-1]
 	*h = old[0 : n-1]
 	return x
+}
+
+func (r *ReconcilePerconaServerMongoDB) isRestoreRunning(cr *api.PerconaServerMongoDB) (bool, error) {
+	restores := api.PerconaServerMongoDBRestoreList{}
+	if err := r.client.List(context.TODO(), &restores, &client.ListOptions{Namespace: cr.Namespace}); err != nil {
+		if k8sErrors.IsNotFound(err) {
+			return false, nil
+		}
+
+		return false, errors.Wrap(err, "get restore list")
+	}
+
+	for _, rst := range restores.Items {
+		if rst.Status.State != api.RestoreStateReady &&
+			rst.Status.State != api.RestoreStateError {
+			return true, nil
+		}
+	}
+
+	return false, nil
+}
+
+func (r *ReconcilePerconaServerMongoDB) isBackupRunning(cr *api.PerconaServerMongoDB) (bool, error) {
+	bcps := api.PerconaServerMongoDBBackupList{}
+	if err := r.client.List(context.TODO(), &bcps, &client.ListOptions{Namespace: cr.Namespace}); err != nil {
+		if k8sErrors.IsNotFound(err) {
+			return false, nil
+		}
+		return false, errors.Wrap(err, "get backup list")
+	}
+
+	for _, bcp := range bcps.Items {
+		if bcp.Status.State != api.BackupStateReady &&
+			bcp.Status.State != api.BackupStateError {
+			return true, nil
+		}
+	}
+
+	return false, nil
 }
