+++ pkg/controller/perconaservermongodb/psmdb_controller.go
@@ -425,6 +425,11 @@ func (r *ReconcilePerconaServerMongoDB) Reconcile(request reconcile.Request) (re
 		}
 	}
 
+	err = r.doInCaseOfRestore(cr)
+	if err != nil {
+		return reconcile.Result{}, errors.Wrap(err, "on restore")
+	}
+
 	err = r.reconcileMongos(cr, mongosTemplateAnnotations)
 	if err != nil {
 		return reconcile.Result{}, errors.Wrap(err, "reconcile mongos")
@@ -601,11 +606,34 @@ func (r *ReconcilePerconaServerMongoDB) deleteCfgIfNeeded(cr *api.PerconaServerM
 	return nil
 }
 
-func (r *ReconcilePerconaServerMongoDB) deleteMongosIfNeeded(cr *api.PerconaServerMongoDB) error {
-	if cr.Spec.Sharding.Enabled {
+func (r *ReconcilePerconaServerMongoDB) doInCaseOfRestore(cr *api.PerconaServerMongoDB) error {
+	if !cr.Spec.Sharding.Enabled {
 		return nil
 	}
 
+	rstRunning, err := r.isRestoreRunning(cr)
+	if err != nil {
+		return errors.Wrap(err, "failed to check running restores")
+	}
+
+	if !rstRunning {
+		return nil
+	}
+
+	err = r.disableBalancerIfNeeded(cr)
+	if err != nil {
+		return errors.Wrap(err, "failed to disable balancer")
+	}
+
+	err = r.deleteMongos(cr)
+	if err != nil {
+		return errors.Wrap(err, "failed to delete mongos")
+	}
+
+	return nil
+}
+
+func (r *ReconcilePerconaServerMongoDB) deleteMongos(cr *api.PerconaServerMongoDB) error {
 	msDepl := psmdb.MongosDeployment(cr)
 	err := r.client.Delete(context.TODO(), msDepl)
 	if err != nil && !k8serrors.IsNotFound(err) {
@@ -621,6 +649,14 @@ func (r *ReconcilePerconaServerMongoDB) deleteMongosIfNeeded(cr *api.PerconaServ
 	return nil
 }
 
+func (r *ReconcilePerconaServerMongoDB) deleteMongosIfNeeded(cr *api.PerconaServerMongoDB) error {
+	if cr.Spec.Sharding.Enabled {
+		return nil
+	}
+
+	return r.deleteMongos(cr)
+}
+
 func (r *ReconcilePerconaServerMongoDB) reconcileMongos(cr *api.PerconaServerMongoDB, secretAnnotations map[string]string) error {
 	if !cr.Spec.Sharding.Enabled {
 		return nil
@@ -631,7 +667,12 @@ func (r *ReconcilePerconaServerMongoDB) reconcileMongos(cr *api.PerconaServerMon
 		return errors.Wrap(err, "failed to chaeck if all sfs are up to date")
 	}
 
-	if !uptodate {
+	rstRunning, err := r.isRestoreRunning(cr)
+	if err != nil {
+		return errors.Wrap(err, "failed to check running restores")
+	}
+
+	if !uptodate || rstRunning {
 		return nil
 	}
